backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
openai>=1.60.0
anthropic>=0.25.0
google-generativeai>=0.5.0
google-genai>=0.7.0
python-dotenv==1.0.1
pydantic==2.6.3
httpx>=0.27.0
